Centralize RSC license header verification
