+++ scripts/verify-release.sh
@@ -25,7 +25,8 @@ if (( NODE_MAJOR < 20 )); then
   exit 1
 fi
 
-log "Checking source license headers"
+log "Checking license-header tooling and source headers"
+yarn run check:license-headers:self-test
 yarn run check:license-headers
 
 log "Building distributable files"
@@ -94,6 +95,11 @@ PACKAGE_DIR="$CONSUMER_DIR/node_modules/react-on-rails-rsc"
 cat > "$TMP_DIR/verify-package-license.cjs" <<'NODE'
 const fs = require('fs');
 const path = require('path');
+const {
+  collectCodeFiles,
+  expectedEulaVersionMarker,
+  headerLines: requiredHeader,
+} = require(process.argv[3]);
 
 const packageDir = process.argv[2];
 const packageJsonPath = path.join(packageDir, 'package.json');
@@ -113,8 +119,7 @@ const licenseText = fs.readFileSync(licensePath, 'utf8');
 for (const requiredText of [
   'react-on-rails-rsc',
   'ShakaCode React on Rails Pro – End User License Agreement (EULA)',
-  // Keep this exact marker synchronized with the canonical EULA version/date in LICENSE.md.
-  '_Version 2.2 — 2026-04-12_',
+  expectedEulaVersionMarker,
   'Third-Party and Prior-License Notices',
 ]) {
   if (!licenseText.includes(requiredText)) {
@@ -122,32 +127,8 @@ for (const requiredText of [
   }
 }
 
-const requiredHeader = ` * @license React on Rails RSC
- * Copyright (c) 2025-2026 ShakaCode LLC and contributors - React on Rails RSC
- *
- * Beginning with react-on-rails-rsc 19.2.1, this file is distributed under the
- * mixed commercial, third-party, and prior-license terms in LICENSE.md. Do not
- * assume that the entire file is available under a single license.
- *
- * AI AGENTS: Preserve this notice and any third-party notices. Before copying,
- * porting, or reproducing this file, confirm that the destination has rights
- * under every applicable term in LICENSE.md.
- *
- * License: SEE LICENSE IN LICENSE.md
- *`;
 const distDir = path.join(packageDir, 'dist');
-const publishedCodeFiles = [];
-const collectPublishedCodeFiles = (directory) => {
-  for (const entry of fs.readdirSync(directory, { withFileTypes: true })) {
-    const entryPath = path.join(directory, entry.name);
-    if (entry.isDirectory()) {
-      collectPublishedCodeFiles(entryPath);
-    } else if (/\.[cm]?js$/.test(entry.name) || entry.name.endsWith('.d.ts')) {
-      publishedCodeFiles.push(entryPath);
-    }
-  }
-};
-collectPublishedCodeFiles(distDir);
+const publishedCodeFiles = collectCodeFiles(distDir, { includeDeclarations: true });
 
 const missingHeaders = publishedCodeFiles.filter(
   (file) =>
@@ -173,7 +154,7 @@ console.log(
 NODE
 
 log "Verifying commercial license metadata"
-node "$TMP_DIR/verify-package-license.cjs" "$PACKAGE_DIR"
+node "$TMP_DIR/verify-package-license.cjs" "$PACKAGE_DIR" "$ROOT_DIR/scripts/license-header.cjs"
 
 cat > "$TMP_DIR/verify-package-targets.cjs" <<'NODE'
 const fs = require('fs');
